fixed stuff moar

+++ penguino-arduino/arduino-patch/scripts/install_avrdude.sh
@@ -1,3 +1,5 @@
 #!/bin/sh
 cd /Applications/Arduino.app/Contents/Resources/Java/hardware/tools/avr/bin
-mv avrdude avrdude-real
+if ! [ -e avrdude-real ]; then
+	mv avrdude avrdude-real;
+fi
